Adds a 'velocity' param to loadFry(), which adds an 'unspliced' assay matrix.
If you want to run RNA velocity on your scRNAseq data, you will need to fetch
the unspliced gene count estimates. The data loading step in loadFry() can add
this data "free of charge" without having to load the results twice, so this
patch allows you to get those counts while you're retrieving the rest of the
data, anyway.

We are assuming you are running avelin-fry / USA quantiation, as described
in this series of tutorials:

https://combine-lab.github.io/alevin-fry-tutorials/tags/#alevin-fry